feat: Implement base repository interfaces and classes
Implement ITournamentRepository and IGameRepository in Core and TournamentRepository and GameRepository in Data.

I change the void methods to return a Task since I call directly into SaveChangesAsync. These should have a way to send commands directly to the database and not await a result; this should be implemented later. On the other hand, it would also be good to ensure that the commands succeed, which we can't do without some database I/O.

I also change GetAsync to return a nullable instead of a regular entity. Since we're making an API it seems reasonable that a user may request an unused id, so I don't want this to be an error.