feat: Change controller classes to use DTO instead of Entity classes
Change the Controller classes to use the new DTO classes instead of Entities. In most cases a direct find-replace works; in some cases the original Entity class still needs to be used.

All the Tournament functions still work, but the Games do not because the TournamentID is missing. In addition, in some cases we would want the Id shown anyway for further navigation. This will be fixed in the next commit.