feat: Add filtering
Add options to filter Games and Tournaments on their Title (using the LINQ case-insensitive Contains).

The only other filters I could think of were on date or time. I feel like to do those justice I would need to implement On, Before, After, and Between. I would like to do that with the specification pattern but I think that's too advanced for this assignment.